refactor page table code, pulling common bits into src/kernel/page.c
